// cat-launcher/src/generated-types/ReleasesUpdatePayload.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GameRelease } from "./GameRelease";
import type { GameVariant } from "./GameVariant";
import type { ReleasesUpdateStatus } from "./ReleasesUpdateStatus";

export type ReleasesUpdatePayload = { variant: GameVariant, releases: Array<GameRelease>, status: ReleasesUpdateStatus, };